BOM Fetching by ID done, Bom creation issue pending Product fetch, Global exception Handling done

+++ BackendAPI/Controllers/ProductController.cs
@@ -1,9 +1,6 @@
-using BackendAPI.Models;
 using BackendAPI.Services;
 using Microsoft.AspNetCore.Mvc;
-using BackendAPI.Dtos;
-using AutoMapper;
-using Microsoft.AspNetCore.Http.HttpResults;
+using BackendAPI.Dtos.Product;
 using BackendAPI.Exceptions;
 
 namespace BackendAPI.Controllers;
@@ -45,17 +42,18 @@ public class ProductsController : ControllerBase
     [HttpPut("{id}")]
     public async Task<IActionResult> Update([FromRoute] int id, [FromBody] ProductUpdateDto dto)
     {
-        Console.WriteLine(id.ToString(), dto);
         if (!ModelState.IsValid)
         return BadRequest(ModelState);
         await _service.UpdateAsync(id,dto);
-        return NoContent();
+        return Ok("Product updated successfully");
     }   
 
     [HttpDelete("{id}")]
     public async Task<IActionResult> Delete(int id)
     {
+        var product = await _service.GetByIdAsync(id);
+        if (product == null) throw new NotFoundException("Product not found");
         await _service.DeleteAsync(id);
-        return NoContent();
+        return Ok("Product deleted successfully");
     }
 }
